+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 1.0.0
+
+- The `verify()` function no longer `throws`.
+- Significantly better test coverage.
+
 # 0.7.0
 
 - Changes naming of `Option` to `LayoutStyle`
@@ -7,7 +12,7 @@
 
 # 0.6.0
 
-- Standardises filenames (which is a breaking change for people already using PixelTest)
+- Standardises filenames (which is a breaking change for people already using PixelTest).
 
 # 0.5.0